chore(channel): standardize PRN data type to usize
- Change PRN field type from i32 to usize in Channel struct
- Remove redundant as usize conversions in signal generator
- Update codegen and allocation logic to use usize consistently

+++ crates/gps/src/channel.rs
@@ -8,7 +8,7 @@ use crate::{
 #[derive(Debug)]
 pub struct Channel {
     /// PRN Number(Pseudorandom Noise)
-    pub prn: i32,
+    pub prn: usize,
     /// C/A Sequence
     pub ca: [i32; CA_SEQ_LEN],
     /// Carrier frequency
